requirements.txt
langchain 
langgraph 
langchain-google-genai
sqlalchemy 
psycopg2-binary
pandas
faiss-cpu
sentence-transformers